[feature] 设备同步相关 1、 修改 data，device 表 ，为其增加 user_id 键 2、 修改 DataBaseOperator，为所有设备操作提供接口 3、 修改所有设备操作相关 Activity，使用新的接口 4、 服务器添加 deviceSyncLet

+++ client/Android/app/src/main/java/com/example/luming/iis/database/DatabaseOperator.java
@@ -48,19 +48,29 @@ public class DatabaseOperator {
                 new Object[]{device.getName(), device.getIp(), device.getPort()});
         System.out.println("addDevice SQL: " + sql);
         db.execSQL(sql);
-        addOperation(sql);
+        addDeviceOperation("-1",sql);
     }
 
-    public void deleteDevice(Device device) {
-        String sql = String.format("delete from device where name = '%s'", new Object[]{device.getName()});
+    public void addDevice(Device device , String user_id) {
+        String sql = String.format("insert into device(user_id,name,ip,port) values( %d,'%s','%s',%s)",
+                new Object[]{user_id,device.getName(), device.getIp(), device.getPort()});
+        System.out.println("addDevice SQL: " + sql);
         db.execSQL(sql);
-        addOperation(sql);
+        addDeviceOperation(user_id,sql);
     }
-    public void updateDevice(Device oldDev, Device newDev){
-        String sql = String.format("update device set name = '%s', ip = '%s', port = %s where name = %s"
-                ,newDev.getName(),newDev.getIp(),newDev.getPort(),oldDev.getName());
+
+    public void deleteDevice(String user_id,Device device) {
+        String sql = String.format("delete from device where name = '%s' and user_id = '%s'", new Object[]{device.getName(),user_id});
+        Log.d("luming","delete sql" + sql);
         db.execSQL(sql);
-        addOperation(sql);
+        addDeviceOperation(user_id,sql);
+    }
+
+    public void updateDevice(String user_id, Device oldDev, Device newDev){
+        String sql = String.format("update device set name = '%s', ip = '%s', port = '%s' where name = '%s' and user_id = '%s'"
+                ,newDev.getName(),newDev.getIp(),newDev.getPort(),oldDev.getName(),user_id);
+        db.execSQL(sql);
+        addDeviceOperation(user_id,sql);
     }
 
     public List<Device> queryAllDevice() {
@@ -68,7 +78,7 @@ public class DatabaseOperator {
         if (!db.isOpen()){
             db = dbHelper.getWritableDatabase();
         }
-        Cursor c = db.rawQuery("select name,ip,port from device", null);
+        Cursor c = db.rawQuery("select name,ip,port from device where user_id = '-1'", null);
         while (c.moveToNext()) {
             Device device = new Device(c.getString(0), c.getString(1), c.getInt(2));
             list.add(device);
@@ -76,6 +86,21 @@ public class DatabaseOperator {
         c.close();
         return list;
     }
+
+    public List<Device> queryAllDevice(Integer user_id) {
+        ArrayList<Device> list = new ArrayList<Device>();
+        if (!db.isOpen()){
+            db = dbHelper.getWritableDatabase();
+        }
+        Cursor c = db.rawQuery("select name,ip,port from device where user_id = " + user_id.toString(),null);
+        while (c.moveToNext()) {
+            Device device = new Device(c.getString(0), c.getString(1), c.getInt(2));
+            list.add(device);
+        }
+        c.close();
+        return list;
+    }
+
 
     /**
      * 根据唯一的name判断是否存在该设备
@@ -90,8 +115,8 @@ public class DatabaseOperator {
         return false;
     }
 
-    public void addLog(String name, String send, String value) {
-        db.execSQL("insert into data (module_name,send,value) values(?,?,?)", new Object[]{name, send, value});
+    public void addData(String user_id, String name, String send, String value) {
+        db.execSQL("insert into data (user_id,module_name,send,value) values(?,?,?)", new Object[]{user_id,name, send, value});
     }
 
     public void addFromJsonArray(String jsonArrayStr) {
@@ -131,20 +156,20 @@ public class DatabaseOperator {
         return list;
     }
 
-    public String queryRecentTime() {
-        Cursor c = db.rawQuery("select time from data order by time desc limit 1", null);
+    public String queryRecentTime(String user_id) {
+        Cursor c = db.rawQuery("select time from data order by time desc limit 1 where user_id = " + user_id, null);
         while (c.moveToNext()) {
             return c.getString(0);
         }
         return "SP_NULL";
     }
 
-    public String queryRecentData(String start) {
+    public String queryRecentData(String start, String user_id) {
         String sql;
-        if (start.equals("SP_NULL"))
-            sql = "select * from data";
+        if (start.equals("NULL"))
+            sql = "select * from data where user_id = " + user_id;
         else
-            sql = "select * from data where time > \"" + start + "\"";
+            sql = "select * from data where time > \"" + start + "\" and user_id =  " + user_id;
         Log.e("sql", sql);
         Cursor c = db.rawQuery(sql, null);
         int num = c.getColumnCount();
@@ -164,7 +189,18 @@ public class DatabaseOperator {
         return array.toString();
     }
 
-    public void addOperation(String sql){
-        db.execSQL("insert into operation(sql) values(?)",new Object[]{sql});
+    public void addDeviceOperation(String user_id,String sql){
+        db.execSQL("insert into operation(user_id,sql) values(?,?)",new Object[]{user_id,sql});
+    }
+
+    public static String getDeviceOperation(String user_id){
+        Cursor cr = db.rawQuery("select sql from operation  where user_id = "+ user_id+ "limit 1",null);
+        if (cr.getCount() != 0){
+            return cr.getString(0);
+        }
+        return null;
+    }
+    public static void popDeviceOperation(String user_id){
+        db.rawQuery("delete from operation where 1 and user_id = "+ user_id + " order by id limit 1",null);
     }
 }
